Trait safe_poll::SafeFuture

pub trait SafeFuture {
    type Output;
    fn safe_poll(
        self: Pin<&mut Self>, 
        context: &mut Context<'_>
    ) -> SafePoll<Self::Output>;
}

An equivalent to Future which returns SafePoll instead of Poll.

Returning SafePoll enforces an extra constraint where a previously created WakeupRegisteredToken must be used. This extra constraint helps to verify that the SafeFuture was properly registered for a wakeup in order to be polled again later. It's possible to fulfill that constraint either by having an internal type return SafePoll::Pending with a token that can be reused or by manually creating the token inside an unsafe block.

Associated Types

type Output

The type of value produced on completion.

Equivalent to Future::Output.

Loading content...
Required methods

fn safe_poll(
self: Pin<&mut Self>,
context: &mut Context<'_>
) -> SafePoll<Self::Output>

Attempt to resolve the asynchronous computation to a final value, registering the current task for wakeup if the value is not yet available.

Equivalent to Future::poll, except it returns a SafePoll to help ensure that the current task was correctly registered for a wakeup.

Loading content...